requirements.txt
flask
flask-cors
langchain-core
langchain-google-genai
chromadb
python-dotenv
sentence_transformers